Skip initializer if `ActualDbSchema` is undefined.
It is recommended to add `actual_db_schema` to the `development` group of your Gemfile. This means that Rails will only `require "actual_db_schema"` when RAILS_ENV is development.

Since initializers run regardless of the Rails environment, the default initializer template causes a `NameError` in non-development environments (the gem is not required and thus the constant is not defined).

```
NameError: uninitialized constant ActualDbSchema (NameError)

  ActualDbSchema.configure do |config|
  ^^^^^^^^^^^^^^
```

To fix this, we only run the ActualDbSchema configure block when `ActualDbSchema` is defined.