English grammar glow-up for README.md (#159)
I _think_ you meant "current" rather than "actual" in this sentence.  In some (many?) languages, the word "actual" (or maybe some cognate of "actual") has the same meaning as the English word "current".  (Spanish and Portuguese are two languages I know of where this is true, and I suspect it's true in many others. French? Italian?  Romanian?  Sardinian?  Catalan?)

In English "actual" has a meaning closer to "real" or "not in error", rather than "current".  (I suspect this is because English borrowed heavily from romance languages—especially French and Latin—but allowed the meaning to drift considerably from the original source.)

What are the implications of this for the name of the project?  
Well, "real" and "not in error" still fits well in the name of the project, so _I am not suggesting any change there_.  This one sentence is the only location of the word "actual" I could find that confused me when I first read it.